Bari_api/requirements.txt
fastapi==0.108.0
uvicorn==0.23.2
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
tensorflow==2.14.0
opencv-python==4.9.0.76
pillow==10.0.0
joblib==1.3.2